% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_normalized_ratio}
\alias{calculate_normalized_ratio}
\title{Calculate cell scores based on number of genes}
\usage{
calculate_normalized_ratio(vec)
}
\arguments{
\item{vec}{list of genes of cell types.}
}
\value{
vector of cell scores based on the number of overlapped genes with the input matrix.
}
\description{
This Function is used to calculate cell scores based on number of genes
}
\author{
Mohmed Soudy \email{Mohamed.soudy@uni.lu} and Sohpie LE BARS \email{sophie.lebars@uni.lu} and Enrico Glaab \email{enrico.glaab@uni.lu}
}
