% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctype.R
\name{correct_gene_symbols}
\alias{correct_gene_symbols}
\title{Process the cell markers names}
\usage{
correct_gene_symbols(markers)
}
\arguments{
\item{markers}{list of unique cell markers.}
}
\value{
vector of genes names which overlap with the correlation matrix.
}
\description{
This Function is used to return the cell markers names processed for the sctype approach
}
\author{
Mohmed Soudy \email{Mohamed.soudy@uni.lu} and Sohpie LE BARS \email{sophie.lebars@uni.lu} and Enrico Glaab \email{enrico.glaab@uni.lu}
}
