% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_correlation}
\alias{filter_correlation}
\title{Filter the genes based on specific correlation threshold}
\usage{
filter_correlation(cor_mat, gene_list, threshold = 0.7)
}
\arguments{
\item{cor_mat}{correlation matrix generated from calculate_cor_mat function.}

\item{gene_list}{cell markers that passed threshold.}

\item{threshold}{absolute correlation threshold.}
}
\value{
vector of gene names that pass user-defined correlation threshold.
}
\description{
This Function is used to filter the gene correlation matrix based on user-defined threshold
}
\author{
Mohmed Soudy \email{Mohamed.soudy@uni.lu} and Sohpie LE BARS \email{sophie.lebars@uni.lu} and Enrico Glaab \email{enrico.glaab@uni.lu}
}
