% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_list}
\alias{filter_list}
\title{Process the cell markers that overlap between the cell markers and scRNA matrix}
\usage{
filter_list(gene_list, passed_cells)
}
\arguments{
\item{gene_list}{list of unique genes of cell types.}

\item{passed_cells}{cells types that pass the specified threshold.}
}
\value{
list of cell types which genes are found in the input matrix.
}
\description{
This Function is used to return the cell markers that overlap between the cell markers and scRNA matrix
}
\author{
Mohmed Soudy \email{Mohamed.soudy@uni.lu} and Sohpie LE BARS \email{sophie.lebars@uni.lu} and Enrico Glaab \email{enrico.glaab@uni.lu}
}
