% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{fisher_test}
\alias{fisher_test}
\title{Performs fisher exact test to get the significant overlap between genes for cell type assignment}
\usage{
fisher_test(ref, gene_overlap)
}
\arguments{
\item{ref}{reference gene set.}

\item{gene_overlap}{genes that pass the correlation threshold.}
}
\value{
vector of p-value and overlap.
}
\description{
This Function is used to perform fisher exact test to get cell types
}
\examples{
fisher_test(c("PAX8","PAX6","TP53","AOC3","LIPF"), c("LIPF","PAX8","PAX6","TP53","TSHB","AOC3"))

}
\author{
Mohmed Soudy \email{Mohamed.soudy@uni.lu} and Sohpie LE BARS \email{sophie.lebars@uni.lu} and Enrico Glaab \email{enrico.glaab@uni.lu}
}
