% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_characters}
\alias{match_characters}
\title{Process the cell markers that pass specific threshold in the gene correlation matrix}
\usage{
match_characters(genes, gene_mat)
}
\arguments{
\item{genes}{list of unique genes of cell types.}

\item{gene_mat}{correlation matrix of genes.}
}
\value{
vector of genes names which overlap with the correlation matrix.
}
\description{
This Function is used to return the cell markers that pass specific threshold in the gene correlation matrix
}
\author{
Mohmed Soudy \email{Mohamed.soudy@uni.lu} and Sohpie LE BARS \email{sophie.lebars@uni.lu} and Enrico Glaab \email{enrico.glaab@uni.lu}
}
