% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{process_clus}
\alias{process_clus}
\title{Gets the associated cell types using correlation-based approach}
\usage{
process_clus(cluster,sobj,assay="RNA",clus,markers,cor_m,m_t=0.9,c_t=0.7,test="p")
}
\arguments{
\item{cluster}{associated cluster name.}

\item{sobj}{Seurat object.}

\item{assay}{assay to be used default is set to RNA.}

\item{clus}{cell clusters.}

\item{markers}{cell markers database.}

\item{cor_m}{gene correlation matrix.}

\item{m_t}{overlap threshold between cell markers and expression matrix.}

\item{c_t}{correlation threshold between genes.}

\item{test}{statistical test that check if overlap is significant could be "p" for phyper or "f" for fisher.}
}
\value{
data frame of proposed cell types.
}
\description{
This Function is used to get the associated cell clusters using correlation-based approach
}
\author{
Mohmed Soudy \email{Mohamed.soudy@uni.lu} and Sohpie LE BARS \email{sophie.lebars@uni.lu} and Enrico Glaab \email{enrico.glaab@uni.lu}
}
