% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctype.R
\name{process_database}
\alias{process_database}
\title{Process the database for the sctype approach}
\usage{
process_database(database_name = "sctype", org = 'a', tissue, tissue_type = 'n')
}
\arguments{
\item{database_name}{name of the database to be used that can be 'sctype' or 'UMD'.}

\item{org}{name of organism to be used that can be 'h' for human, 'm' for mouse, and 'a' for all markers.}

\item{tissue}{specified tissue from which the data comes.}

\item{tissue_type}{tissue type whether 'a' for all types 'n' for normal tissues only or "c" for cancer tissues.}
}
\value{
vector of genes names which overlap with the correlation matrix.
}
\description{
This Function is used to process the database that will be used for sctype approach
}
\author{
Mohmed Soudy \email{Mohamed.soudy@uni.lu} and Sohpie LE BARS \email{sophie.lebars@uni.lu} and Enrico Glaab \email{enrico.glaab@uni.lu}
}
