% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_markers}
\alias{process_markers}
\title{Process the cell markers database and return the processed list}
\usage{
process_markers(markers_df)
}
\arguments{
\item{markers_df}{data frame with markers named as gene_original and cell names as cell type.}
}
\value{
list of lists of the processed markers
}
\description{
This Function is used to process the cell markers database and return the processed list
}
\author{
Mohmed Soudy \email{Mohamed.soudy@uni.lu} and Sohpie LE BARS \email{sophie.lebars@uni.lu} and Enrico Glaab \email{enrico.glaab@uni.lu}
}
