% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{sccca}
\alias{sccca}
\title{Run the pipeline for the cell type assignment}
\usage{
sccca(sobj,assay="RNA",cluster,marker,tissue,tt="a",cond,m_t=0.9,c_t=0.7,test="p",org="a")
}
\arguments{
\item{sobj}{Seurat object.}

\item{assay}{assay to be used default is set to RNA.}

\item{cluster}{colname in the mata.data that have the cell cluster numbers.}

\item{marker}{cell markers database path.}

\item{tissue}{specified tissue from which the data comes.}

\item{tt}{tissue type whether 'a' for all types 'n' for normal tissues only or "c" for cancer tissues.}

\item{cond}{colname in the meta.data that have the condition names.}

\item{m_t}{overlap threshold between cell markers and expression matrix.}

\item{c_t}{correlation threshold between genes.}

\item{test}{statistical test that check if overlap is significant could be "p" for phyper or "f" for fisher.}

\item{org}{organism to be used that can be 'h' for human, 'm' for mouse, and 'a' for all markers.}
}
\value{
list of Seurat object that have the assigned clusters, and top 3 proposed cell types.
}
\description{
This Function is used to run the main pipeline that does the cell type assignment
}
\author{
Mohmed Soudy \email{Mohamed.soudy@uni.lu} and Sohpie LE BARS \email{sophie.lebars@uni.lu} and Enrico Glaab \email{enrico.glaab@uni.lu}
}
