% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctype.R
\name{sctype}
\alias{sctype}
\title{Run the sctype approach as it's implemented by Ianevski, A., Giri, A.K. and Aittokallio, T.}
\usage{
sctype(sobj,assay="RNA",tissue,tt="a",clus,org="a",scaled=T,database="sctype")
}
\arguments{
\item{sobj}{Seurat object.}

\item{assay}{assay to be used default is set to RNA.}

\item{tissue}{specified tissue from which the data comes.}

\item{tt}{tissue type whether 'a' for all types 'n' for normal tissues only or "c" for cancer tissues.}

\item{clus}{colname in the mata.data that have the cell cluster numbers.}

\item{org}{organism to be used that can be 'h' for human, 'm' for mouse, and 'a' for all markers.}

\item{scaled}{indicates whether the matrix is scaled (TRUE by default)}

\item{database}{name of the database to be used that can be 'sctype' or 'UMD'}
}
\value{
vector of genes names which overlap with the correlation matrix.
}
\description{
This Function is used to run the sctype approach with faster implementation
}
\author{
Mohmed Soudy \email{Mohamed.soudy@uni.lu} and Sohpie LE BARS \email{sophie.lebars@uni.lu} and Enrico Glaab \email{enrico.glaab@uni.lu}
}
